+++ ExtendedTooltip/Patches/SystemOrderPatches.cs
@@ -16,7 +16,8 @@ namespace ExtendedTooltip.Patches
             updateSystem?.UpdateAt<CustomTranslationSystem>(SystemUpdatePhase.UIUpdate);
             updateSystem?.UpdateAt<ExtendedTooltipUISystem>(SystemUpdatePhase.UIUpdate);
             updateSystem?.UpdateAt<TempElevationTooltipSystem>(SystemUpdatePhase.UITooltip);
-            updateSystem?.UpdateBefore<ExtendedTooltipSystem, RaycastNameTooltipSystem>(SystemUpdatePhase.UITooltip);
+            updateSystem?.UpdateAt<ExtendedTooltipSystem>(SystemUpdatePhase.UITooltip);
+            updateSystem?.UpdateAt<ExtendedBulldozerTooltipSystem>(SystemUpdatePhase.UITooltip);
         }
     }
 }